+++ example/lib/main.dart
@@ -264,7 +264,7 @@ class _MyAppState extends State<MyApp> {
                 fontSize: 15,
               ),
             ),
-            !isExpanded ? () => controller.scrollTo(230) : controller.collapse,
+            !isExpanded ? () => controller.scrollTo(state.maxScrollExtent) : controller.collapse,
             color: Colors.white,
             border: BorderSide(
               color: Colors.grey.shade400,
@@ -487,6 +487,7 @@ class _MyAppState extends State<MyApp> {
           duration: const Duration(milliseconds: 500),
           snapSpec: const SnapSpec(
             snap: true,
+            initialSnap: 1.0,
             snappings: [
               0.3,
               0.7,